app/requirements.txt
langchain>=0.1.0
langchain-google-vertexai>=0.1.0
langgraph>=0.1.0
pydantic>=2.0.0
streamlit>=1.28.0
nest-asyncio>=1.5.0 
langchain-experimental==0.3.4
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0